requirements/base.txt
Django==4.1.7
djangorestframework==3.14.0
hashids==1.3.1